Rename to facade
