+++ lint-tasks.yml
@@ -4,10 +4,10 @@ vars:
   G_LINT_VENV_DIR: "{{.BUILD_DIR}}/lint-venv"
   G_LINT_VENV_CHECKSUM_FILE: "{{.BUILD_DIR}}/lint#venv.md5"
   # Linter target dirs
-  G_CPP_WRAPPED_DEPENDENCY_HEADERS_DIR: "{{.ROOT_DIR}}/src/wrapped_dependency_headers"
+  G_CPP_WRAPPED_FACADE_HEADERS_DIR: "{{.ROOT_DIR}}/src/wrapped_facade_headers"
   G_CPP_LINT_DIRS:
     - "{{.CLP_FFI_PY_CPP_SRC_DIR}}"
-    - "{{.G_CPP_WRAPPED_DEPENDENCY_HEADERS_DIR}}"
+    - "{{.G_CPP_WRAPPED_FACADE_HEADERS_DIR}}"
   G_PYTHON_LINT_DIRS: ["{{.ROOT_DIR}}/clp_ffi_py", "{{.ROOT_DIR}}/tests"]
 
 tasks:
@@ -111,7 +111,7 @@ tasks:
             # and so that the workflow doesn't fail due to violations in other files.
             - "{{.CLP_FFI_PY_CPP_SRC_DIR}}/Py_utils.cpp"
             - "{{.CLP_FFI_PY_CPP_SRC_DIR}}/Py_utils.hpp"
-            - "{{.G_CPP_WRAPPED_DEPENDENCY_HEADERS_DIR}}"
+            - "{{.G_CPP_WRAPPED_FACADE_HEADERS_DIR}}"
           VENV_DIR: "{{.G_LINT_VENV_DIR}}"
 
   py-check:
