+++ src/clp_ffi_py/.clang-format
@@ -4,7 +4,7 @@ IndentExternBlock: "NoIndent"
 IncludeCategories:
   # NOTE: A header is grouped by first matching regex
   # Python header must be included before any others
-  - Regex: "^<wrapped_dependency_headers/Python\\.hpp>$"
+  - Regex: "^<wrapped_facade_headers/Python\\.hpp>$"
     Priority: -1
   # Project headers
   - Regex: "^<clp_ffi_py"
@@ -14,7 +14,7 @@ IncludeCategories:
     Priority: 3
   # Third-party headers with wrappers to silent clang-tidy warnings. Update when adding new
   # third-party library
-  - Regex: "<wrapped_dependency_headers"
+  - Regex: "<wrapped_facade_headers"
     Priority: 3
   # C headers
   - Regex: "^<.*.h>$"
