+++ src/clp_ffi_py/PyObjectCast.hpp
@@ -1,7 +1,7 @@
 #ifndef CLP_FFI_PY_PY_OBJECT_CAST_HPP
 #define CLP_FFI_PY_PY_OBJECT_CAST_HPP
 
-#include <wrapped_dependency_headers/Python.hpp>  // Must be included before any other header files
+#include <wrapped_facade_headers/Python.hpp>  // Must be included before any other header files
 
 #include <type_traits>
