+++ src/clp_ffi_py/Py_utils.cpp
@@ -1,4 +1,4 @@
-#include <wrapped_dependency_headers/Python.hpp>  // Must be included before any other header file
+#include <wrapped_facade_headers/Python.hpp>  // Must be included before any other header file
 
 #include "Py_utils.hpp"
