+++ src/clp_ffi_py/ir/native/PyKeyValuePairLogEvent.cpp
@@ -1,4 +1,4 @@
-#include <wrapped_dependency_headers/Python.hpp>  // Must be included before any other header files
+#include <wrapped_facade_headers/Python.hpp>  // Must be included before any other header files
 
 #include "PyKeyValuePairLogEvent.hpp"
