+++ src/clp_ffi_py/ir/native/PyMetadata.cpp
@@ -1,4 +1,4 @@
-#include <wrapped_dependency_headers/Python.hpp>
+#include <wrapped_facade_headers/Python.hpp>
 
 #include "PyMetadata.hpp"
