+++ src/clp_ffi_py/ir/native/PyQuery.cpp
@@ -1,5 +1,5 @@
 
-#include <wrapped_dependency_headers/Python.hpp>  // Must be included before any other header files
+#include <wrapped_facade_headers/Python.hpp>  // Must be included before any other header files
 
 #include "PyQuery.hpp"
