+++ src/clp_ffi_py/ir/native/deserialization_methods.cpp
@@ -1,4 +1,4 @@
-#include <wrapped_dependency_headers/Python.hpp>  // Must be included before any other header files
+#include <wrapped_facade_headers/Python.hpp>  // Must be included before any other header files
 
 #include "deserialization_methods.hpp"
