+++ src/clp_ffi_py/utils.cpp
@@ -1,4 +1,4 @@
-#include <wrapped_dependency_headers/Python.hpp>  // Must be included before any other header files
+#include <wrapped_facade_headers/Python.hpp>  // Must be included before any other header files
 
 #include "utils.hpp"
 
@@ -8,7 +8,7 @@
 
 #include <clp/TraceableException.hpp>
 #include <outcome/single-header/outcome.hpp>
-#include <wrapped_dependency_headers/msgpack.hpp>
+#include <wrapped_facade_headers/msgpack.hpp>
 
 #include <clp_ffi_py/ExceptionFFI.hpp>
 #include <clp_ffi_py/PyObjectCast.hpp>
